test(hubspot): move association id regression test
